Moving to diskimage-builder
This patch makes it possible to build the Virtualbox-compatible OVA
archive withouth using vagrant, virtualbox, or vmware-specific tools.

Currently the created ova archive can be imported on VirtualBox and
VMWare Player, the disk image itself can be used with libvirt
(qemu-kvm). It contains all the drivers that enables using it
as a desktop system with shared clipboard, auto-resizing display, etc.

Since the diskimage-builder method installs in chroot, there is no way
to pre-install devstack by running stack.sh. The image makes a best
effort including all the deb and pip packages to make the first run of
stack.sh as short as possible.

Change-Id: Ie1f6e44db7e0af8d072cfdf16ace1ba3f749c246