Allow users to override output format in create-training-box.sh
Users may want to override the generated output format, in case they
want to run in a virtualization environment other than VMware or
VirtualBox.

Accept a positional argument enabling users to select their preferred
output format, defaulting to VMDK (the previously hard-coded format).

Also, make the creation of an OVA archive conditional on the user
selecting a disk format commonly used with VirtualBox (that is, VDI or
VMDK).

Change-Id: I8829c1a21aa79de877748d9bdc3b7eef7802a6d8